fixed bug of bug
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Retreat2008@4111 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/DATASTRUCTURES/SuffixArray.h
@@ -77,7 +77,7 @@ public:
 	@throw InvalidValue if the spectrum is not sorted ascendingly
 	
 	*/
-	virtual void findSpec(std::vector<std::vector<std::pair<std::pair<int, int>,float > > >& candidates, const std::vector<double> & spec) = 0;
+	virtual void findSpec(std::vector<std::vector<std::pair<std::pair<int, int>,double > > >& candidates, const std::vector<double> & spec) = 0;
 
 	/**
 	@brief saves the suffix array to disc
